+++ packages/react-query/src/internals/getClientArgs.ts
@@ -1,0 +1,7 @@
+export function getClientArgs<TPathAndInput extends unknown[], TOptions>(
+  pathAndInput: TPathAndInput,
+  opts: TOptions,
+) {
+  const [path, input] = pathAndInput;
+  return [path, input, (opts as any)?.trpc] as const;
+}
